fix: numberOfCycles type
Fixed type issue for numberOfCycles attemting to cast string to integer

+++ service/grails-app/controllers/org/olf/PredictedPieceSetController.groovy
@@ -46,7 +46,7 @@ class PredictedPieceSetController extends OkapiTenantAwareController<PredictedPi
 
     ArrayList<UserConfiguredTemplateMetadata> startingValues = new ArrayList<UserConfiguredTemplateMetadata>(startingValuesJson)
 
-    Integer numberOfCycles = data?.numberOfCycles ?: 1
+    Integer numberOfCycles = data?.numberOfCycles as Integer ?: 1
 
     ArrayList<InternalPiece> ips = pieceGenerationService.createPiecesTransient(ruleset, LocalDate.parse(data.startDate), numberOfCycles)
